auth-service/requirements.txt
anyio==3.6.1
astroid==2.11.7
asyncpg==0.26.0
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
colorama==0.4.5
databases==0.6.0
dill==0.3.5.1
ecdsa==0.18.0
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
loguru==0.6.0
mccabe==0.7.0
passlib==1.7.4
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pylint==2.14.5
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
tomlkit==0.11.2
typing_extensions==4.3.0
uvicorn==0.18.2
win32-setctime==1.1.0
wrapt==1.14.1
